requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.3.1
django-grappelli==4.0.1
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
kombu==5.3.7
msgpack==1.0.8
mysql-connector-python==8.4.0
mysqlclient==2.2.4
pillow==10.3.0
prompt_toolkit==3.0.45
PyJWT==2.8.0
pymongo==4.7.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
